Add .pro file and replace special characters

+++ src/controlwidget.cpp
@@ -1,5 +1,5 @@
 /*
-*  Copyright 2020 José María Castelo Ares
+*  Copyright 2020 Jose Maria Castelo Ares
 *
 *  Contact: <jose.maria.castelo@gmail.com>
 *  Repository: <https://github.com/jmcastelo/MorphogenGL>
@@ -680,7 +680,7 @@ void ControlWidget::initPipelineControls(int selectedPipelineIndex)
 {
     // Remove all pipeline controls
 
-    for (int pipelineIndex = pipelinePushButton.size() - 1; pipelineIndex >= 0; pipelineIndex--)
+    for (int pipelineIndex = static_cast<int>(pipelinePushButton.size()) - 1; pipelineIndex >= 0; pipelineIndex--)
     {
         pipelinesButtonGroup->removeButton(pipelinePushButton[pipelineIndex]);
         pipelinesGridLayout->removeWidget(pipelinePushButton[pipelineIndex]);
