+++ src/pipeline.h
@@ -1,5 +1,5 @@
 /*
-*  Copyright 2020 José María Castelo Ares
+*  Copyright 2020 Jose Maria Castelo Ares
 *
 *  Contact: <jose.maria.castelo@gmail.com>
 *  Repository: <https://github.com/jmcastelo/MorphogenGL>
@@ -57,7 +57,7 @@ public:
         std::vector<float> matrixElements);
 
     QString getImageOperationName(int operationIndex) { return imageOperations[operationIndex]->getName(); };
-    int getImageOperationsSize() { return imageOperations.size(); };
+    int getImageOperationsSize() { return static_cast<int>(imageOperations.size()); };
 
     GLuint getTextureID();
